Removed unnecessary component
